# Improve performance of Random method by not using "arguments"