@nok 2016-07-12
This thread results in a similar result and recommendation: http://stackoverflow.com/a/13299598/1293700 
I am also for it, because the performance of the core library matters.
Here we find the relevant parts: https://github.com/processing/p5.js/search?utf8=%E2%9C%93&q=arguments