@rwaldron 2016-07-12
@kadamwhite asked me to review this change, but since I am not an active participant I will keep my notes as general as possible.