@rwaldron 2016-07-12
Two notes: 

1. `instanceof` will fail across realms, so if there is any possibility that `random(...)` might be called with input from an iframe or called in iframe with input from a parent frame, then it will fail. With a minor perf hit on check for the cross-realm arrays, changing the condition to the following will resolve the issue:

  ```js
  if (min instanceof Array || Array.isArray(min)) {
  ```
  ...However it might just be that this is a use case that's not intended to be supported, in which case: freely ignore :)

2. It might be worthwhile to add some kind of grunt-task benchmark testing, @shama's [`grunt-benchmark`](https://github.com/shama/grunt-benchmark) seems like a good place to start :)