@rwaldron 2016-07-12
One more thing! Is your CI not hooked up to pull requests?