@kadamwhite 2016-07-14
100% in favor of wiring up grunt-benchmark; added a note to #1524.

I don't think the iframe realm question is cause for concern, but that's a @lmccart _et al_ question. Same with the final Q; I assume PR's be triggering CI, but don't have the full context there.