@lmccart 2016-07-15
Yeah, I don't think iframe realm crossing is a major use case so this seems fine to me. PRs do trigger CI, but it takes a moment to run, so a PR may show up initially as mergeable if there are no conflicts and later display an error. (I'm pretty sure? Is there something you saw that seems like it's not working properly?)